back-end/docker_test/requirement.txt
pymongo==3.11.3
pyshark==0.4.3
flask==1.1.1
ipcalc==1.99.0
psutil==5.8.0
get_nic==0.0.3
six==1.15.0
python-dateutil==2.6.1
Flask-PyMongo==2.3.0
